# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indian Cinema Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movies:
    get:
      operationId: listMovies
      tags: [movies]
      summary: List movies in library
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [watched, watchlist]
        - name: language
          in: query
          schema:
            type: string
        - name: rating
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
    post:
      operationId: createMovie
      tags: [movies]
      summary: Add a movie to library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
...
  /movies/{id}:
    get:
      operationId: getMovie
      tags: [movies]
      summary: Get a movie by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Not found
    patch:
      operationId: updateMovie
      tags: [movies]
      summary: Update a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieUpdate"
      responses:
        "200":
          description: Updated movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Not found
    delete:
      operationId: deleteMovie
      tags: [movies]
      summary: Delete a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /movies/{id}/match:
    post:
      operationId: matchMovieToTmdb
      tags: [movies]
      summary: Match a single movie to TMDB
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated movie with TMDB data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Not found
  /movies/{id}/rewatch:
    post:
      operationId: rewatchMovie
      tags: [movies]
      summary: Log a rewatch for an already-watched movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewatchInput"
      responses:
        "200":
          description: Updated movie after rewatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "400":
          description: Movie is not marked watched
        "404":
          description: Not found
...
components:
  schemas:
...
    Movie:
      type: object
      required: [id, title, status, createdAt, rewatchCount, rewatchDates]
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [watched, watchlist]
        rating:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tmdbId:
          type: ["integer", "null"]
        posterPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        releaseDate:
          type: ["string", "null"]
          description: Theatrical/digital release day (YYYY-MM-DD) for upcoming reminders
        originalLanguage:
          type: ["string", "null"]
        genres:
          type: ["array", "null"]
          items:
            type: string
        overview:
          type: ["string", "null"]
        watchedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        rewatchCount:
          type: integer
          minimum: 0
          description: Number of rewatches after the first watch (0 = watched once)
        rewatchDates:
          type: array
          items:
            type: string
          description: >
            Optional dates logged for rewatches (ISO 8601). Length may be less
            than rewatchCount when some rewatches were logged without a date.
    MovieInput:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
          minLength: 1
        status:
          type: string
          enum: [watched, watchlist]
        rating:
          type: string
        notes:
          type: string
        tmdbId:
          type: integer
        posterPath:
          type: string
        releaseYear:
          type: integer
        releaseDate:
          type: string
          description: Theatrical/digital release day (YYYY-MM-DD)
        originalLanguage:
          type: string
        genres:
          type: array
          items:
            type: string
        overview:
          type: string
        watchedAt:
          type: string
    MovieUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        status:
          type: string
          enum: [watched, watchlist]
        rating:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tmdbId:
          type: ["integer", "null"]
        posterPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        releaseDate:
          type: ["string", "null"]
          description: Theatrical/digital release day (YYYY-MM-DD)
        originalLanguage:
          type: ["string", "null"]
        genres:
          type: ["array", "null"]
          items:
            type: string
        overview:
          type: ["string", "null"]
        watchedAt:
          type: ["string", "null"]
    RewatchInput:
      type: object
      properties:
        rating:
          type: ["string", "null"]
          description: Optional updated rating from this rewatch
        watchedAt:
          type: ["string", "null"]
          description: >
            Optional date for this rewatch (ISO 8601 date or datetime).
            Omit or null to log a rewatch without recording a date.
...