# lib/api-zod/src/generated/types/movie.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Indian Cinema Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MovieStatus } from './movieStatus';

export interface Movie {
  id: number;
  title: string;
  status: MovieStatus;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tmdbId?: number | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  releaseYear?: number | null;
  /**
     * Theatrical/digital release day (YYYY-MM-DD) for upcoming reminders
     * @nullable
     */
  releaseDate?: string | null;
  /** @nullable */
  originalLanguage?: string | null;
  /** @nullable */
  genres?: string[] | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  watchedAt?: string | null;
  createdAt: string;
  /**
     * Number of rewatches after the first watch (0 = watched once)
     * @minimum 0
     */
  rewatchCount: number;
  /** Optional dates logged for rewatches (ISO 8601). Length may be less than rewatchCount when some rewatches were logged without a date. */
  rewatchDates: string[];
}