# lib/api-zod/src/generated/types/movieInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Indian Cinema Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MovieInputStatus } from './movieInputStatus';

export interface MovieInput {
  /** @minLength 1 */
  title: string;
  status: MovieInputStatus;
  rating?: string;
  notes?: string;
  tmdbId?: number;
  posterPath?: string;
  releaseYear?: number;
  /** Theatrical/digital release day (YYYY-MM-DD) */
  releaseDate?: string;
  originalLanguage?: string;
  genres?: string[];
  overview?: string;
  watchedAt?: string;
}