# lib/api-zod/src/generated/types/movieUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Indian Cinema Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MovieUpdateStatus } from './movieUpdateStatus';

export interface MovieUpdate {
  /** @minLength 1 */
  title?: string;
  status?: MovieUpdateStatus;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tmdbId?: number | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  releaseYear?: number | null;
  /**
     * Theatrical/digital release day (YYYY-MM-DD)
     * @nullable
     */
  releaseDate?: string | null;
  /** @nullable */
  originalLanguage?: string | null;
  /** @nullable */
  genres?: string[] | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  watchedAt?: string | null;
}